Moved all models to Knowledge.Models project
