+++ src/CognitiveSearch.UI/CognitiveSearch.UI/Views/_ViewImports.cshtml
@@ -2,7 +2,7 @@
      Licensed under the MIT License. -->
 
 @using CognitiveSearch.UI
-@using Knowledge.Services.Models
+@using Knowledge.Models
 @addTagHelper *, WebOptimizer.Core
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @inject Microsoft.ApplicationInsights.AspNetCore.JavaScriptSnippet JavaScriptSnippet
