+++ src/CognitiveSearch.UI/Knowledge.API/Controllers/AbstractApiController.cs
@@ -1,9 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Knowledge.Services;
 using Knowledge.Configuration;
-using Knowledge.Services.Models;
+using Knowledge.Models;
+using Knowledge.Services;
 using Microsoft.ApplicationInsights;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
