+++ src/CognitiveSearch.UI/Knowledge.Models/IResultItem.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Knowledge.Services
+namespace Knowledge.Models
 {
     public interface IResultItem
     {
