+++ src/CognitiveSearch.UI/Knowledge.Models/Knowledge.Models.csproj
@@ -7,7 +7,13 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Azure.Search.Documents" Version="11.3.0" />
+    <PackageReference Include="Microsoft.Spatial" Version="7.12.2" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Knowledge.Configuration\Knowledge.Configuration.csproj" />
   </ItemGroup>
 
 </Project>
