+++ src/CognitiveSearch.UI/Knowledge.Models/SearchRequest.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Knowledge.Services.Models
+namespace Knowledge.Models
 {
     using Newtonsoft.Json;
     using Newtonsoft.Json.Serialization;
