+++ src/CognitiveSearch.UI/Knowledge.Services/Answers/QnA/QnAResponse.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Knowledge.Models.Answers;
 using System.Collections.Generic;
 
 namespace Knowledge.Services.QnA
@@ -9,22 +10,5 @@ namespace Knowledge.Services.QnA
     {
         public IList<Answer> answers { get; set; }
         public bool activeLearningEnabled { get; set; }
-    }
-
-    public class Answer: IResultItem 
-    {
-        public string[] questions { get; set; }
-        public string answer { get; set; }
-        public float score { get; set; }
-        public int id { get; set; }
-        public string source { get; set; }
-        public object[] metadata { get; set; }
-        public Context context { get; set; }
-    }
-
-    public class Context
-    {
-        public bool isContextOnly { get; set; }
-        public object[] prompts { get; set; }
     }
 }
