+++ src/CognitiveSearch.UI/Knowledge.Services/AzureSearch/AbstractSearchService.cs
@@ -8,9 +8,9 @@ using Azure.Storage.Blobs;
 using Azure.Storage.Sas;
 using Knowledge.Configuration;
 using Knowledge.Configuration.Graph;
+using Knowledge.Models;
 using Knowledge.Services.AzureStorage;
 using Knowledge.Services.Helpers;
-using Knowledge.Services.Models;
 using System;
 using System.Collections.Generic;
 using System.Net;
