+++ src/CognitiveSearch.UI/Knowledge.Services/AzureSearch/SDK/AzureSearchSDKService.cs
@@ -9,10 +9,9 @@ using Azure.Search.Documents.Models;
 using Knowledge.Configuration;
 using Knowledge.Configuration.Graph;
 using Knowledge.Models;
+using Knowledge.Models.Ingress;
 using Knowledge.Services.AzureStorage;
 using Knowledge.Services.Helpers;
-using Knowledge.Services.Models;
-using Knowledge.Services.Models.Ingress;
 using Knowledge.Services.SemanticSearch;
 using Microsoft.ApplicationInsights;
 using Newtonsoft.Json;
