+++ src/CognitiveSearch.UI/Knowledge.Services/Graph/GraphRequest.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Knowledge.Services.Models.Ingress;
+using Knowledge.Models.Ingress;
 using System.Collections.Generic;
 
 namespace Knowledge.Services.Graph
