+++ src/CognitiveSearch.UI/Knowledge.Services/Helpers/PermissionsHelper.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Knowledge.Services.Models;
+using Knowledge.Models;
 using System;
 using System.Collections.Generic;
