+++ src/CognitiveSearch.UI/Knowledge.Services/IQueryService.cs
@@ -1,8 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Knowledge.Services.Models;
-using Knowledge.Services.Models.Ingress;
+using Knowledge.Models;
+using Knowledge.Models.Ingress;
 using System.Collections.Generic;
 using System.Threading.Tasks;
