+++ src/CognitiveSearch.UI/Knowledge.Services/WebSearch/WebSearchRequest.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Knowledge.Services.Models.Ingress;
+using Knowledge.Models.Ingress;
 
 namespace Knowledge.Services.WebSearch
 {
